refactor runtime code
