+++ WDL/runtime/_workflow_state.py
@@ -1,0 +1,751 @@
+"""
+Workflow state machine and graph helpers.
+"""
+
+import itertools
+import json
+import logging
+import math
+from typing import Any, Callable, Dict, Iterable, List, NamedTuple, Optional, Set, Tuple, Union
+
+import regex
+
+from .. import Env, Error, StdLib, Tree, Type, Value
+from .._util import WDLVersion, wdl_version_geq
+from .._util import StructuredLogMessage as _
+from . import config
+from ._io_helpers import (
+    _fspaths,
+    _resolve_source_relative_paths,
+    _resolve_workflow_path,
+    _source_directory,
+)
+
+
+class WorkflowOutputs(Tree.WorkflowNode):
+    """
+    A no-op workflow node which depends on each ``Decl`` node from the workflow output section. Or,
+    if the workflow is missing the output section, depends on ``Call`` and ``Gather`` nodes for all
+    call outputs.
+
+    The workflow state machine tacks this on to the workflow graph to facilitate assembly of the
+    outputs environment.
+    """
+
+    output_node_ids: Set[str]
+
+    def __init__(self, workflow: Tree.Workflow) -> None:
+        super().__init__("outputs", workflow.pos)
+
+        self.output_node_ids = set()
+        if workflow.outputs is not None:
+            for node in workflow.outputs:
+                self.output_node_ids.add(node.workflow_node_id)
+        else:
+            # no output{} section -- use all top-level Calls and any top-level Gather whose
+            # ultimate referee is a Call
+            for n in workflow.body:
+                if isinstance(n, Tree.Call):
+                    self.output_node_ids.add(n.workflow_node_id)
+                if isinstance(n, Tree.WorkflowSection):
+                    for g in n.gathers.values():
+                        if isinstance(g.final_referee, Tree.Call):
+                            self.output_node_ids.add(g.workflow_node_id)
+
+    def _workflow_node_dependencies(self) -> Iterable[str]:
+        yield from self.output_node_ids
+
+    def add_to_type_env(
+        self, struct_types: Env.Bindings[Dict[str, Type.Base]], type_env: Env.Bindings[Type.Base]
+    ) -> Env.Bindings[Type.Base]:
+        raise NotImplementedError()
+
+    @property
+    def children(self) -> Iterable[Tree.SourceNode]:
+        return []
+
+
+_Job = NamedTuple(
+    "_Job",
+    [
+        ("id", str),
+        ("node", Tree.WorkflowNode),
+        ("dependencies", Set[str]),
+        ("scatter_stack", List[Tuple[str, Env.Binding[Value.Base]]]),
+    ],
+)
+
+
+class StateMachine:
+    """
+    On-line workflow state machine, suitable for use within a singleton driver process managing
+    in-memory state. The state machine evaluates WDL expressions locally, while instructing the
+    driver when to call tasks/subworkflows. It's agnostic to how/where the driver actually executes
+    each call, just requiring asynchronous notification of call completion along with the outputs.
+    """
+
+    _logger: Optional[logging.Logger] = None
+    logger_id: str
+    run_dir: str
+    values_to_json: Callable[[Env.Bindings[Value.Base]], Dict]
+    workflow: Tree.Workflow
+    inputs: Env.Bindings[Value.Base]
+    jobs: Dict[str, _Job]
+    job_outputs: Dict[str, Env.Bindings[Value.Base]]
+    finished: Set[str]
+    running: Set[str]
+    waiting: Set[str]
+
+    # File/Directory paths that were either expressly supplied as workflow inputs, or were
+    # generated by this workflow execution. By default, these are the only local paths the workflow
+    # is allowed to access. (Unless [file_io] allow_any_input = true)
+    fspath_allowlist: Set[str]
+    # TODO: factor out WorkflowState interface?
+
+    def __init__(
+        self,
+        logger_id: str,
+        run_dir: str,
+        workflow: Tree.Workflow,
+        inputs: Env.Bindings[Value.Base],
+    ) -> None:
+        """
+        Initialize the workflow state machine from the workflow AST and inputs
+        """
+        self.logger_id = logger_id
+        self.run_dir = run_dir
+        self.workflow = workflow
+        self.inputs = inputs
+        self.jobs = {}
+        self.job_outputs = {}
+        self.finished = set()
+        self.running = set()
+        self.waiting = set()
+        self.fspath_allowlist = _fspaths(inputs)
+
+        from .. import values_to_json
+
+        self.values_to_json = values_to_json  # type: ignore
+
+        # Preprocess inputs: if None value is supplied for an input declared with a default but
+        # without the ? type quantifier, remove the binding entirely so that the default will be
+        # used. In contrast, if the input declaration has an -explicitly- optional type, then we'll
+        # allow the supplied None to override any default.
+        input_decls = workflow.available_inputs
+        self.inputs = self.inputs.filter(
+            lambda b: (
+                not (
+                    isinstance(b.value, Value.Null)
+                    and b.name in input_decls
+                    and input_decls[b.name].expr
+                    and not input_decls[b.name].type.optional
+                )
+            )
+        )
+
+        workflow_nodes = (workflow.inputs or []) + workflow.body + (workflow.outputs or [])
+        workflow_nodes.append(WorkflowOutputs(workflow))
+
+        for node in workflow_nodes:
+            deps = node.workflow_node_dependencies
+            if isinstance(node, Tree.Decl):
+                # strike the dependencies of any decl node whose value is supplied in the inputs
+                if self.inputs.has_binding(node.name):
+                    deps = set()
+            self._schedule(
+                _Job(id=node.workflow_node_id, node=node, dependencies=deps, scatter_stack=[])
+            )
+
+        # TODO: by topsorting all section bodies we could ensure that when we schedule an
+        # additional job, all its dependencies will already have been scheduled, increasing
+        # flexibility/compatibility with various backends.
+
+        # sanity check
+        assert "outputs" in self.jobs
+        known_jobs = set(self.waiting)
+        for node in workflow_nodes:
+            if isinstance(node, Tree.WorkflowSection):
+                for gather in node.gathers.values():
+                    known_jobs.add(gather.workflow_node_id)
+        for job in self.jobs.values():
+            assert not (job.dependencies - known_jobs), (
+                job.id,
+                (job.dependencies - known_jobs),
+                known_jobs,
+            )
+
+    @property
+    def outputs(self) -> Optional[Env.Bindings[Value.Base]]:
+        """
+        Workflow outputs, once the workflow is completely finished. ``None`` until then.
+
+        Warning: be sure to distinguish ``None``, the workflow isn't finished, from ``[]``, the
+        workflow finished with no outputs.
+        """
+        if len(self.finished) < len(self.jobs):
+            return None
+        ans = self.job_outputs["outputs"]
+        assert ans is not None
+        return ans
+
+    CallInstructions = NamedTuple(
+        "CallInstructions",
+        [
+            ("id", str),
+            ("callee", Union[Tree.Task, Tree.Workflow]),
+            ("inputs", Env.Bindings[Value.Base]),
+        ],
+    )
+    """
+    The state machine produces a ``CallInstructions`` object when it's time for the driver to
+    launch a task/subworkflow job.
+
+    :param id: call/job ID string, unique in the workflow
+    :param callee: ``WDL.Call`` or ``WDL.Workflow`` to launch
+    :param inputs: ``WDL.Env.Bindings[Value.Base]`` of call inputs
+    """
+
+    def step(
+        self, cfg: config.Loader, stdlib: StdLib.Base
+    ) -> "Optional[StateMachine.CallInstructions]":
+        """
+        Advance the workflow state machine, returning the next call to initiate.
+
+        The driver must start the specified callee task/workflow and then, later upon its
+        completion, invoke ``call_finished()`` with its outputs. It is NOT necessary to await the
+        call's completion before another ``step()`` for the next call; this allows the driver
+        to orchestrate multiple calls at once. Indeed, the driver should launch as many calls as
+        it can support concurrently, by calling ``step()`` in a loop until getting back ``None``;
+        doing so after initialization and after each ``call_finished()`` invocation, until at last
+        the workflow outputs are available.
+        """
+        runnable: List[str] = []
+        while True:
+            # select a job whose dependencies are all finished
+            if not runnable:
+                runnable = sorted(
+                    [j for j in self.waiting if not (self.jobs[j].dependencies - self.finished)],
+                    reverse=True,
+                )
+            if not runnable:
+                assert self.running or not self.waiting, "deadlocked: " + str(
+                    set(itertools.chain(*(self.jobs[j].dependencies for j in self.waiting)))
+                    - self.finished
+                )
+                return None
+            job_id = runnable.pop()
+            job = self.jobs[job_id]
+
+            # mark it 'running'
+            self.running.add(job.id)
+            self.waiting.remove(job.id)
+
+            # do the job
+            try:
+                res = self._do_job(cfg, stdlib, job)
+            except Exception as exn:
+                setattr(exn, "job_id", job.id)
+                raise exn
+
+            # if it's a call, return instructions to the driver
+            if isinstance(res, StateMachine.CallInstructions):
+                return res
+
+            # otherwise, record the outputs, mark the job finished, and move on to the next job
+            envlog = self.values_to_json(res)
+            self.logger.info(
+                _(
+                    "visit",
+                    node=job.id,
+                    values=envlog if len(json.dumps(envlog)) < 4096 else "(((large)))",
+                )
+            )
+            self.job_outputs[job.id] = res
+            self.running.remove(job.id)
+            self.finished.add(job.id)
+
+    def call_finished(self, job_id: str, outputs: Env.Bindings[Value.Base]) -> None:
+        """
+        Deliver notice of a job's successful completion, along with its outputs
+        """
+        assert job_id in self.running
+        outlog = self.values_to_json(outputs)
+        self.logger.notice(_("finish", job=job_id))
+        self.logger.info(
+            _(
+                "output",
+                job=job_id,
+                values=outlog if len(json.dumps(outlog)) < 4096 else "(((large)))",
+            )
+        )
+        call_node = self.jobs[job_id].node
+        assert isinstance(call_node, Tree.Call)
+        self.job_outputs[job_id] = outputs.wrap_namespace(call_node.name)
+        self.fspath_allowlist |= _fspaths(outputs)
+        self.finished.add(job_id)
+        self.running.remove(job_id)
+
+    def _schedule(self, job: _Job) -> None:
+        self.logger.debug(_("schedule", node=job.id, dependencies=list(job.dependencies)))
+        assert job.id not in self.jobs
+        self.jobs[job.id] = job
+        self.waiting.add(job.id)
+
+    def _do_job(
+        self, cfg: config.Loader, stdlib: StdLib.Base, job: _Job
+    ) -> "Union[StateMachine.CallInstructions, Env.Bindings[Value.Base]]":
+        if isinstance(job.node, Tree.Gather):
+            return _gather(
+                job.node, dict((dep_id, self.job_outputs[dep_id]) for dep_id in job.dependencies)
+            )
+
+        # for all non-Gather nodes, derive the environment by merging the outputs of all the
+        # dependencies (+ any current scatter variable bindings)
+        scatter_vars: Env.Bindings[Value.Base] = Env.Bindings()
+        for p in job.scatter_stack:
+            scatter_vars = Env.Bindings(p[1], scatter_vars)
+        env = Env.merge(scatter_vars, *(self.job_outputs[dep] for dep in job.dependencies))
+        envlog = self.values_to_json(env)
+        self.logger.debug(
+            _(
+                "env",
+                node=job.id,
+                values=envlog if len(json.dumps(envlog)) < 4096 else "(((large)))",
+            )
+        )
+
+        if isinstance(job.node, (Tree.Scatter, Tree.Conditional)):
+            for newjob in _scatter(
+                self.workflow,
+                job.node,
+                env,
+                job.scatter_stack,
+                stdlib,
+                cfg.get_int("scheduler", "scatter_tag_max"),
+            ):
+                self._schedule(newjob)
+            # the section node itself has no outputs, so return an empty env
+            return Env.Bindings()
+
+        if isinstance(job.node, Tree.Decl):
+            # bind the value obtained either (i) from the workflow inputs or (ii) by evaluating
+            # the expr
+            v = _eval_decl(cfg, self.inputs, self.fspath_allowlist, job.node, env, stdlib)
+            return Env.Bindings(Env.Binding(job.node.name, v))
+
+        if isinstance(job.node, WorkflowOutputs):
+            return Value.rewrite_env_paths(
+                env,
+                lambda v: _resolve_workflow_path(cfg, self.fspath_allowlist, "workflow output", v),
+            )
+
+        if isinstance(job.node, Tree.Call):
+            # evaluate input expressions
+            call_name = job.node.name
+            call_inputs: Env.Bindings[Value.Base] = Env.Bindings()
+            for name, expr in job.node.inputs.items():
+                call_inputs = call_inputs.bind(name, expr.eval(env, stdlib=stdlib))
+            # check workflow inputs for additional inputs supplied to this call
+            for b in self.inputs.enter_namespace(call_name):
+                call_inputs = call_inputs.bind(b.name, b.value)
+
+            # coerce inputs to required types (treating inputs with defaults as optional even if
+            # they don't have the ? type quantifier)
+            assert isinstance(job.node.callee, (Tree.Task, Tree.Workflow))
+            callee_inputs = job.node.callee.available_inputs
+            call_inputs = _postprocess_call_inputs(
+                cfg,
+                stdlib.wdl_version,
+                self.fspath_allowlist,
+                call_name,
+                callee_inputs,
+                call_inputs,
+                source_directory=_source_directory(self.workflow),
+            )
+            # issue CallInstructions
+            self.logger.notice(_("ready", job=job.id, callee=job.node.callee.name))
+            inplog = self.values_to_json(call_inputs)
+            self.logger.info(
+                _(
+                    "input",
+                    job=job.id,
+                    values=inplog if len(json.dumps(inplog)) < 4096 else "(((large)))",
+                )
+            )
+
+            return StateMachine.CallInstructions(
+                id=job.id, callee=job.node.callee, inputs=call_inputs
+            )
+
+        raise NotImplementedError()
+
+    @property
+    def logger(self) -> logging.Logger:
+        if not self._logger:
+            self._logger = logging.getLogger(self.logger_id)
+            # TODO: if we were truly unpickling in a new process, we'd need to add a new
+            # LoggingFileHandler to self._logger
+        return self._logger
+
+    def __getstate__(self) -> Dict[str, Any]:
+        ans = dict(self.__dict__)
+        del ans["_logger"]  # for Python pre-3.7 loggers: https://bugs.python.org/issue30520
+        return ans
+
+
+def _scatter(
+    workflow: Tree.Workflow,
+    section: Union[Tree.Scatter, Tree.Conditional],
+    env: Env.Bindings[Value.Base],
+    scatter_stack: List[Tuple[str, Env.Binding[Value.Base]]],
+    stdlib: StdLib.Base,
+    max_tag: int,
+) -> Iterable[_Job]:
+    # we'll be tracking, for each body node ID, the IDs of the potentially multiple corresponding
+    # jobs scheduled
+    multiplex: Dict[str, Set[str]] = {}
+    for body_node in section.body:
+        multiplex[body_node.workflow_node_id] = set()
+        if isinstance(body_node, Tree.WorkflowSection):
+            for subgather in body_node.gathers.values():
+                multiplex[subgather.workflow_node_id] = set()
+
+    # evaluate scatter array or boolean condition
+    v = section.expr.eval(env, stdlib=stdlib)
+    array: List[Optional[Value.Base]] = []
+    if isinstance(section, Tree.Scatter):
+        assert isinstance(v, Value.Array)
+        for v_i in v.value:
+            array.append(v_i)
+    else:
+        assert isinstance(v, Value.Boolean)
+        if v.value:
+            # condition is satisfied, so we'll "scatter" over a length-1 array
+            array = [None]
+    digits = math.ceil(math.log10(len(array))) if len(array) > 1 else 1
+
+    # for each array element, schedule an instance of the body subgraph
+    last_scatter_indices = None
+    scatter_tags = _scatter_tags(array, max_tag)
+    for i, array_i in enumerate(array):
+        # scatter bookkeeping: format the index as a left-zero-padded string so that it'll sort
+        # lexicographically in the desired order; bind the scatter variable name to the array value
+        scatter_stack_i = scatter_stack
+        if isinstance(array_i, Value.Base):
+            assert isinstance(section, Tree.Scatter)
+            str_i = str(i).zfill(digits)
+            assert len(str_i) <= digits
+            if scatter_tags[i]:
+                str_i += "-" + scatter_tags[i]
+            scatter_stack_i = scatter_stack_i + [(str_i, Env.Binding(section.variable, array_i))]
+        scatter_indices_i = [p[0] for p in scatter_stack_i]
+        assert last_scatter_indices is None or last_scatter_indices < scatter_indices_i
+        last_scatter_indices = scatter_indices_i
+
+        # schedule each body (template) node
+        for body_node in section.body:
+            # the job ID will be the template node ID with the current scatter index appended.
+            # if we're in nested scatters, then append *each* respective index!
+            assert len(scatter_indices_i) == body_node.scatter_depth
+            body_job_id = _append_scatter_indices(body_node.workflow_node_id, scatter_indices_i)
+
+            # furthermore, rewrite the template node's dependencies on other within-scatter nodes
+            # to the corresponding jobs given the current scatter index.
+            # especially tricky: in a nested scatter we can depend on a node at a higher level, for
+            # which we need to append only the indices up to its level!
+            dependencies = set()
+            for dep_id in body_node.workflow_node_dependencies:
+                dep = workflow.get_node(dep_id)
+                assert dep.scatter_depth <= body_node.scatter_depth
+                dependencies.add(
+                    _append_scatter_indices(dep_id, scatter_indices_i[: dep.scatter_depth])
+                )
+
+            yield _Job(
+                id=body_job_id,
+                node=body_node,
+                dependencies=dependencies,
+                scatter_stack=scatter_stack_i,
+            )
+
+            # record the newly scheduled job & its expected gathers in multiplex
+            multiplex[body_node.workflow_node_id].add(body_job_id)
+            if isinstance(body_node, Tree.WorkflowSection):
+                for subgather in body_node.gathers.values():
+                    multiplex[subgather.workflow_node_id].add(
+                        _append_scatter_indices(subgather.workflow_node_id, scatter_indices_i)
+                    )
+
+    # schedule each gather op with dependencies multiplexed onto the set of jobs scheduled
+    # from the corresponding body node.
+    # if the scatter array was empty or the condition was false, these dependencies are
+    # empty, so these jobs will become runnable immediately to "gather" empty arrays or
+    # Value.Null's as appropriate.
+    for body_node_id, gather in section.gathers.items():
+        yield _Job(
+            id=_append_scatter_indices(gather.workflow_node_id, [p[0] for p in scatter_stack]),
+            node=gather,
+            dependencies=multiplex[body_node_id],
+            scatter_stack=scatter_stack,
+        )
+
+
+def _append_scatter_indices(node_id: str, scatter_indices: List[str]) -> str:
+    return "-".join([node_id] + scatter_indices)
+
+
+def _scatter_tags(array: List[Optional[Value.Base]], max_tag: int) -> List[str]:
+    # Given an array of values, compute an array of names for each item that strives for useful
+    # human-readability, and can be embedded in the run id/directory safely. This is to help the
+    # operator navigate the run logs & directory tree, looking for specific items.
+
+    # stringify each item and split each string into a list of identifier-like components
+    any = False
+    delimiters = regex.compile("[^0-9a-zA-Z_]+")
+    items = []
+    for i, array_i in enumerate(array):
+        if (
+            isinstance(array_i, Value.Base)
+            and not isinstance(array_i, Value.Null)
+            and not (isinstance(array_i, Value.Int) and array_i.value == i)
+        ):
+            items.append(delimiters.split(json.dumps(array_i.json)))
+            any = True
+        else:
+            items.append([])
+    if not any or max_tag <= 0:
+        return [""] * len(array)
+
+    # compute & remove those lists' longest common prefix & suffix
+    lcp = _longest_common_prefix(items)
+    if lcp:
+        items = [item[lcp:] for item in items]
+    lcs = _longest_common_prefix([list(reversed(item)) for item in items])
+    if lcs:
+        items = [item[:-lcs] for item in items]
+
+    # concatenate remaining components
+    tags = ["".join(item) for item in items]
+
+    # truncate to first max_tag characters
+    tags_pfx = [tag[:max_tag].rstrip("-") for tag in tags]
+    if sum(1 for tag in tags_pfx if len(tag)) == sum(1 for tag in set(tags_pfx) if len(tag)):
+        return tags_pfx
+    # if those weren't unique, then try suffix; if those aren't unique either, then give up to
+    # avoid generating misleading tags.
+    tags_sfx = [tag[-max_tag:].lstrip("-") for tag in tags]
+    return (
+        tags_sfx
+        if sum(1 for tag in tags_sfx if len(tag)) == sum(1 for tag in set(tags_sfx) if len(tag))
+        else ([""] * len(items))
+    )
+
+
+def _longest_common_prefix(items: List[List[str]]) -> int:
+    ans = 0
+    for i, item in enumerate(items):
+        if i == 0:
+            ans = len(item)
+        else:
+            ans = min(ans, len(item))
+            prev_item = items[i - 1]
+            j = 0
+            while j < ans and item[j] == prev_item[j]:
+                j += 1
+            ans = j
+        if ans == 0:
+            return 0
+    return ans
+
+
+def _gather(
+    gather: Tree.Gather, dependencies: Dict[str, Env.Bindings[Value.Base]]
+) -> Env.Bindings[Value.Base]:
+    # important: the dependency job IDs must sort lexicographically in the desired array order!
+    dep_ids = sorted(dependencies.keys())
+
+    # since it would be so awful to permute the array silently, lets verify the ID order
+    if len(dep_ids) > 1:
+        assert isinstance(gather.section, Tree.Scatter)
+        dep_ids_split = [dep_id.split("-") for dep_id in dep_ids]
+        dep_id_lcp = _longest_common_prefix(dep_ids_split)
+        for i, dep_id_i in enumerate(dep_ids_split):
+            assert i == int(dep_id_i[dep_id_lcp])
+
+    # figure out names of the values to gather, either the name if the referenced decl,
+    # or each output of the referenced call.
+    leaf = gather.final_referee  # follow potential linked list of Gathers for nested sections
+    if isinstance(leaf, Tree.Decl):
+        names = [leaf.name]
+    elif isinstance(leaf, Tree.Call):
+        names = []
+        outp = leaf.effective_outputs.enter_namespace(leaf.name)
+        assert len(outp) == len(leaf.effective_outputs)
+        for b in outp:
+            names.append(b.name)
+    else:
+        assert False
+
+    # for each such name,
+    ans: Env.Bindings[Value.Base] = Env.Bindings()
+    ns = [leaf.name] if isinstance(leaf, Tree.Call) else []
+    for name in names:
+        # gather the corresponding values
+        values = [dependencies[dep_id].resolve(".".join(ns + [name])) for dep_id in dep_ids]
+        v0 = values[0] if values else None
+        assert v0 is None or isinstance(v0, Value.Base)
+        # bind the array, singleton value, or None as appropriate
+        if isinstance(gather.section, Tree.Scatter):
+            rhs: Value.Base = Value.Array((v0.type if v0 else Type.Any()), values)
+        else:
+            assert isinstance(gather.section, Tree.Conditional)
+            assert len(values) <= 1
+            rhs = v0 if v0 is not None else Value.Null()
+        ans = ans.bind(".".join(ns + [name]), rhs)
+
+    return ans
+
+
+def _eval_decl(
+    cfg: config.Loader,
+    inputs: Env.Bindings[Value.Base],
+    allowlist: Set[str],
+    decl: Tree.Decl,
+    env: Env.Bindings[Value.Base],
+    stdlib: StdLib.Base,
+) -> Value.Base:
+    """
+    Evaluate one workflow declaration and check its File/Directory paths.
+
+    WDL 1.2 source-relative declaration paths are resolved to host paths and allowlisted here.
+    Missing children of allowlisted input Directories are rewritten to Null before final type
+    coercion, so only File?/Directory? declarations accept them.
+
+    ``allowlist`` is intentionally mutated only to add resolved source-relative declaration paths.
+    """
+    if decl.name in inputs:
+        value = inputs[decl.name]
+    elif decl.expr:
+        value = decl.expr.eval(env, stdlib=stdlib).coerce(decl.type)
+        if wdl_version_geq(stdlib.wdl_version, WDLVersion.V1_2):
+            value, source_paths = _resolve_source_relative_paths(
+                cfg,
+                _source_directory(decl),
+                value,
+                decl.type,
+                f"workflow declaration {decl.name}",
+            )
+            allowlist |= source_paths
+    else:
+        assert decl.type.optional
+        value = Value.Null()
+
+    value = Value.rewrite_paths(
+        value,
+        lambda v: _resolve_workflow_path(
+            cfg,
+            allowlist,
+            f"workflow declaration {decl.name}",
+            v,
+            null_if_missing=True,
+        ),
+    )
+    try:
+        return value.coerce(decl.type)
+    except FileNotFoundError:
+        raise Error.InputError(f"File/Directory path not found in workflow declaration {decl.name}")
+
+
+def _postprocess_call_inputs(
+    cfg: config.Loader,
+    wdl_version: str,
+    allowlist: Set[str],
+    call_name: str,
+    callee_inputs: Env.Bindings[Tree.Decl],
+    call_inputs: Env.Bindings[Value.Base],
+    source_directory: str = "",
+) -> Env.Bindings[Value.Base]:
+    """
+    Coerce call inputs, then check their File/Directory paths for use as host paths.
+
+    The first coercion exposes String paths as File/Directory values. After path checks may rewrite
+    missing input Directory children to Null, the second coercion enforces optionality.
+
+    ``source_directory`` should be the local directory containing the workflow source file, or ""
+    when unavailable. WDL 1.2+ relative call-input paths are resolved against it before the
+    allowlist checks.
+    """
+    call_inputs = _coerce_call_inputs(callee_inputs, call_inputs)
+    if wdl_version_geq(wdl_version, WDLVersion.V1_2):
+        call_inputs, source_paths = _resolve_call_input_source_paths(
+            cfg, source_directory, call_name, callee_inputs, call_inputs
+        )
+        allowlist |= source_paths
+    call_inputs = Value.rewrite_env_paths(
+        call_inputs,
+        lambda v: _resolve_workflow_path(
+            cfg,
+            allowlist,
+            f"call {call_name} input",
+            v,
+            null_if_missing=True,
+        ),
+    )
+    try:
+        return _coerce_call_inputs(callee_inputs, call_inputs)
+    except FileNotFoundError:
+        raise Error.InputError(f"File/Directory path not found in call {call_name} input")
+
+
+def _coerce_call_inputs(
+    callee_inputs: Env.Bindings[Tree.Decl], call_inputs: Env.Bindings[Value.Base]
+) -> Env.Bindings[Value.Base]:
+    """
+    Coerce call input values to their callee declaration types.
+
+    Declarations with defaults are treated as optional here so Null can still select the default
+    during task/subworkflow input evaluation.
+    """
+
+    def coerce_binding(binding: Env.Binding[Value.Base]) -> Env.Binding[Value.Base]:
+        if binding.name not in callee_inputs:
+            return binding
+        callee_decl = callee_inputs[binding.name]
+        callee_type = callee_decl.type.copy(optional=True) if callee_decl.expr else callee_decl.type
+        return Env.Binding(binding.name, binding.value.coerce(callee_type))
+
+    return call_inputs.map(coerce_binding)
+
+
+def _resolve_call_input_source_paths(
+    cfg: config.Loader,
+    source_directory: str,
+    call_name: str,
+    callee_inputs: Env.Bindings[Tree.Decl],
+    call_inputs: Env.Bindings[Value.Base],
+) -> Tuple[Env.Bindings[Value.Base], Set[str]]:
+    """
+    Resolve WDL 1.2 source-relative paths supplied directly to call inputs.
+
+    The values are already coerced to callee input types. This second pass uses the caller workflow
+    source directory, not the callee declaration source directory, and returns the source paths to
+    add to the workflow allowlist. ``source_paths`` is mutated locally while walking bindings; no
+    arguments are mutated.
+    """
+    source_paths: Set[str] = set()
+
+    def resolve_binding(binding: Env.Binding[Value.Base]) -> Env.Binding[Value.Base]:
+        callee_decl = callee_inputs.get(binding.name)
+        if not callee_decl:
+            return binding
+        value, paths = _resolve_source_relative_paths(
+            cfg,
+            source_directory,
+            binding.value,
+            callee_decl.type.copy(optional=True) if callee_decl.expr else callee_decl.type,
+            f"call {call_name} input {binding.name}",
+        )
+        source_paths.update(paths)
+        return Env.Binding(binding.name, value, binding.info)
+
+    return call_inputs.map(resolve_binding), source_paths
